working Makefile (still everything newly compiled, when 1 c changes), ft_strtrim

+++ ft_strlcat.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lbaumann < lbaumann@student.42berlin.de    +#+  +:+       +#+        */
+/*   By: lbaumann <lbaumann@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 13:28:52 by lbaumann          #+#    #+#             */
-/*   Updated: 2022/11/30 11:13:45 by lbaumann         ###   ########.fr       */
+/*   Updated: 2022/12/05 11:42:57 by lbaumann         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	dst_len;
-	int	src_len;
+	size_t	dst_len;
+	size_t	src_len;
 
 	dst_len = ft_strlen(dst);
 	src_len = ft_strlen(src);
